+++ src/compas_timber/fabrication/btlx.py
@@ -194,8 +194,8 @@ class BTLxWriter(object):
         parts_element = ET.SubElement(project_element, "Parts")
         # create part elements for each beam
         elements = chain(model.beams, model.plates)
-        for element in elements:
-            part_element = self._create_part(element, element.graphnode)
+        for i, element in enumerate(elements):
+            part_element = self._create_part(element, order_num=i)
             parts_element.append(part_element)
         return project_element
 
@@ -365,15 +365,18 @@ class BTLxGenericPart(object):
         The width of the part.
     height : float
         The height of the part.
+    name : str, optional
+        The name of the part. Defaults to an empty string.
     scale_factor : float, optional
         The scale factor to apply to the part's dimensions. Defaults to 1.0.
     """
 
-    def __init__(self, order_num, length, width, height, scale_factor=1.0):
+    def __init__(self, order_num, length, width, height, name="", scale_factor=1.0):
         self.order_num = order_num
         self.length = length * scale_factor
         self.width = width * scale_factor
         self.height = height * scale_factor
+        self.name = name
         self._scale_factor = scale_factor
 
     @property
@@ -431,7 +434,7 @@ class BTLxGenericPart(object):
             "AssemblyNumber": "",
             "OrderNumber": str(self.order_num),
             "Designation": "",
-            "Annotation": "",
+            "Annotation": str(self.name) + "-" + str(self.part_guid)[:4],
             "Storey": "",
             "Group": "",
             "Package": "",
@@ -445,7 +448,7 @@ class BTLxGenericPart(object):
             "Weight": "0",
             "ProcessingQuality": "automatic",
             "StoreyType": "",
-            "ElementNumber": "00",
+            "ElementNumber": str(self.part_guid)[:4],
             "Layer": "0",
             "ModuleNumber": "",
         }
@@ -487,6 +490,7 @@ class BTLxRawpart(BTLxGenericPart):
             stock.length,
             stock.width,
             stock.height,
+            stock.name,
             scale_factor,
         )
         self.stock = stock
@@ -577,7 +581,7 @@ class BTLxPart(BTLxGenericPart):
     """
 
     def __init__(self, element, order_num, scale_factor=1.0):
-        super(BTLxPart, self).__init__(order_num, element.blank_length, element.width, element.height, scale_factor)
+        super(BTLxPart, self).__init__(order_num, element.blank_length, element.width, element.height, element.name, scale_factor)
         self.element = element
         self.processings = []
         self._shape_strings = None